// graphql/generated/fields/task.fields.ts
/* Do not edit this file. It was generated programmatically. */

const fields = {
  title: {
    label: "title",
    type: "String",
    required: true,
    widget: {},
  },
  description: {
    label: "description",
    type: "String",
    required: false,
    widget: { multiline: true, fullWidth: true },
  },
  plannedStartDate: {
    label: "plannedStartDate",
    type: "DateTime",
    required: false,
    widget: {},
  },
  dueDate: {
    label: "dueDate",
    type: "DateTime",
    required: false,
    widget: {},
  },
  userId: {
    label: "userId",
    type: "ID",
    required: true,
    widget: {},
  },
  parentId: {
    label: "parentId",
    type: "ID",
    required: false,
    widget: {},
  },
  habitId: {
    label: "habitId",
    type: "ID",
    required: false,
    widget: {},
  },
  rank: {
    label: "rank",
    type: "Number",
    required: true,
    widget: {},
  },
  completedAt: {
    label: "completedAt",
    type: "DateTime",
    required: false,
    widget: {},
  },
};

export default fields;