// graphql/generated/fragments/task.fragment.ts
/* Do not edit this file. It was generated programmatically. */

import { Task } from "@/graphql/generated/models/task.model";
import { gql } from "@apollo/client";

export const taskFragment = gql`
  fragment TaskFragment on Task {
    __typename
    id
    title
    description
    plannedStartDate
    dueDate
    userId
    parentId
    habitId
    rank
    completedAt
    createdAt
    updatedAt
    archivedAt
  }
`;

export type TaskFragment = Pick<
  Task,
  | "__typename"
  | "id"
  | "createdAt"
  | "updatedAt"
  | "archivedAt"
  | "title"
  | "description"
  | "plannedStartDate"
  | "dueDate"
  | "userId"
  | "parentId"
  | "habitId"
  | "rank"
  | "completedAt"
>;